Create UI.py and switched updated UI code to it
