+++ miniproj7.py
@@ -1,258 +1,187 @@
-import sys
-from PyQt6.QtWidgets import QApplication, QWidget, QLabel, QLineEdit, QPushButton, QTextEdit, QVBoxLayout, QHBoxLayout, QMessageBox, QGridLayout
-from PyQt6.QtGui import QPixmap, QFont
-from PyQt6.QtCore import Qt
-from textblob import TextBlob
+import tkinter as tk
+from spellchecker import SpellChecker as sc
 from nltk.corpus import wordnet
 import pyttsx3
 
-# Initialize text-to-speech engine
-eng = pyttsx3.init()
+from tkinter import PhotoImage
+from PIL import ImageTk,Image
 
-# User dictionary and history
-user_dict = set()
+from tkinter import ttk
+import customtkinter
+from tkinter import messagebox
+
+
+spell_checker = sc()
+eng = pyttsx3.init()
 history_tracking = []
 
-class PhraseCraftApp(QWidget):
-    def __init__(self):
-        super().__init__()
 
-        self.init_ui()
 
-    def init_ui(self):
-        self.setWindowTitle("PhraseCraft v0.5 - Updated")
-        self.setGeometry(100, 100, 1000, 700)
+def spellcheck():
+	text=entry1.get()
+	history_tracking.append(text)
+	misspelled = spell_checker.unknown(text.lower().split())
+	corrected_text = text
 
-        self.setStyleSheet("""
-            QWidget {
-                background-color: #1C1C1C; 
-                font-family: 'Segoe Print'; 
-            }
-        """)
+	for word in misspelled:
+		suggestions = spell_checker.correction(word)
+		if suggestions != word:
+			question = f"Is '{suggestions}' the correct spelling for '{word}'? (Y/N)"
+			user_response = messagebox.askquestion("Spelling Check", question)
+			if user_response == "yes":
+				corrected_text = corrected_text.replace(word, suggestions)
+				entry1.delete(0,"end")
+				entry1.insert(10, corrected_text)
+			else:
+				suggestions_text = f"Suggestions for {word}: {', '.join(spell_checker.candidates(word))}"
+				messagebox.showinfo("Spelling Suggestions", suggestions_text)
 
-        layout = QVBoxLayout()
+def pronun():
+	word=entry1.get()
+	eng.say(word)
+	eng.runAndWait()
 
-        self.create_app_bar(layout)
+def get_word_info():
+	word = entry1.get()
+	synsets = wordnet.synsets(word)
+	meanings = [syn.definition() for syn in synsets] if synsets else []
 
-        self.create_search_area(layout)
+	output= {
+		"word": word,
+		"meaning": meanings,
+		"origin": get_word_origin(word),
+	}
+	global label_out
+	label_out=tk.Label(root,text=output,font=("georgia",24))
+	label_out.place(x=2,y=600)
 
-        self.create_action_buttons(layout)
 
-        self.create_output_area(layout)
+def get_word_origin(word):
+	synsets = wordnet.synsets(word)
+	og = set()
+	for syn in synsets:
+		for lemma in syn.lemmas():
+			og.add(lemma.name())
+	return list(og)
 
-        self.setLayout(layout)
 
-    def create_app_bar(self, layout):
-        appbar_layout = QHBoxLayout()
+def get_synonyms():
+	word = entry1.get()
+	synsets = wordnet.synsets(word)
+	synonyms = set()
+	for syn in synsets:
+		for lemma in syn.lemmas():
+			synonyms.add(lemma.name())
 
-        title_label = QLabel("PhraseCraft", self)
-        title_label.setFont(QFont("Arial", 32, QFont.Weight.Bold))
-        title_label.setStyleSheet("color: #FFEA00;")
-        title_label.setAlignment(Qt.AlignmentFlag.AlignCenter)
-        appbar_layout.addWidget(title_label)
+	op=list(synonyms)[:10]
+	global label_synonym
+	label_synonym=tk.Label(master=root,text=op,font=("georgia",24))
+	label_synonym.place(x=2,y=900)
+	
 
-        appbar_layout.addStretch(1)
-        layout.addLayout(appbar_layout)
+user_dict = set()
 
-    def create_search_area(self, layout):
-        search_layout = QHBoxLayout()
+def track_history(word):
+	history_tracking.insert(0, word)
+	if len(history_tracking) > 10:
+		history_tracking = history_tracking[:10]
 
-        self.text_input = QLineEdit(self)
-        self.text_input.setPlaceholderText("Enter text here...")
-        self.text_input.setFont(QFont("Arial", 18))
-        self.text_input.setFixedHeight(60)
-        self.text_input.setStyleSheet("""
-            QLineEdit {
-                border-radius: 20px;
-                padding: 15px;
-                border: 2px solid #FFEA00;
-                background-color: #2A2A2A;  
-                color: white;
-            }
-            QLineEdit:focus {
-                border: 2px solid #76b852;
-            }
-        """)
-        search_layout.addWidget(self.text_input)
+def add_to_user_dict():
+	word= entry1.get()
+	user_dict.add(word)
 
-        self.btn_spellcheck = QPushButton("Spellcheck", self)
-        self.btn_spellcheck.setFont(QFont("Arial", 20))
-        self.btn_spellcheck.setFixedHeight(60)
-        self.btn_spellcheck.setStyleSheet("""
-            QPushButton {
-                border-radius: 20px;
-                padding: 15px;
-                background-color: #FFEA00;  
-                color: #000000;
-            }
-            QPushButton:hover {
-                background-color: #FFD700;  
-            }
-            QPushButton:pressed {
-                background-color: #FFC107;  
-            }
-        """)
-        self.btn_spellcheck.clicked.connect(self.spellcheck)
-        search_layout.addWidget(self.btn_spellcheck)
+def disp_ud():
+	global label_ud
+	label_ud=tk.Label(master=root,text=list(user_dict)[:],font=("georgia",24))
+	label_ud.place(x=2,y=700)
+def remove_from_dic():
+	word = entry1.get()
+	user_dict.discard(word)
+def disp_history():
+	global label_hist
+	label_hist=tk.Label(master=root,text=history_tracking[:10],font=("georgia",24))
+	label_hist.place(x=2,y=800)
 
-        layout.addLayout(search_layout)
+def remove():
+	if "label_ud" in globals():
+		label_ud.place_forget()
+	if "label_out" in globals():
+		label_out.place_forget()
+	if "label_synonym" in globals():
+		label_synonym.place_forget()
+	if "label_hist" in globals():
+		label_hist.place_forget()
 
-    def create_action_buttons(self, layout):
-        buttons_layout = QGridLayout()
+def update_colors():
+	bg_color = "#1c1c1e" if dark_mode_var.get() else "#fff8c9"
+	text_color = "white" if dark_mode_var.get() else "black"
 
-        self.btn_pronounce = QPushButton("Pronounce", self)
-        self.set_button_style(self.btn_pronounce)
-        self.btn_pronounce.clicked.connect(self.pronounce_word)
-        buttons_layout.addWidget(self.btn_pronounce, 0, 0)
+	root.configure(bg=bg_color)
+	mode_button.configure(bg=bg_color, fg=text_color)
+	
 
-        self.btn_meaning_origin = QPushButton("Get Meanings and Origin", self)
-        self.set_button_style(self.btn_meaning_origin)
-        self.btn_meaning_origin.clicked.connect(self.get_word_info)
-        buttons_layout.addWidget(self.btn_meaning_origin, 0, 1)
+def toggle_mode():
+	current_mode = dark_mode_var.get()
+	new_mode = not current_mode
+	dark_mode_var.set(new_mode)
+	mode_text = "Light Mode" if new_mode else "Dark Mode"
+	mode_button.configure(text=mode_text)
+	update_colors()
 
-        self.btn_synonyms = QPushButton("Get Synonyms", self)
-        self.set_button_style(self.btn_synonyms)
-        self.btn_synonyms.clicked.connect(self.get_synonyms)
-        buttons_layout.addWidget(self.btn_synonyms, 1, 0)
+	
+root=tk.Tk()
+root.title("PhraseCraft v0.5-Presentable")
+root.geometry("1000x1000")
+#
+speaker=ImageTk.PhotoImage(Image.open(r"/home/Suchitra/Desktop/code./dictionary/speaker.jpeg"))
 
-        self.btn_add_dict = QPushButton("Add to My Dictionary", self)
-        self.set_button_style(self.btn_add_dict)
-        self.btn_add_dict.clicked.connect(self.add_to_user_dict)
-        buttons_layout.addWidget(self.btn_add_dict, 1, 1)
 
-        self.btn_display_dict = QPushButton("Display My Dictionary", self)
-        self.set_button_style(self.btn_display_dict)
-        self.btn_display_dict.clicked.connect(self.display_user_dict)
-        buttons_layout.addWidget(self.btn_display_dict, 2, 0)
+dark_mode_var = tk.BooleanVar()
+dark_mode_var.set(False)
 
-        self.btn_remove_dict = QPushButton("Remove from my Dictionary", self)
-        self.set_button_style(self.btn_remove_dict)
-        self.btn_remove_dict.clicked.connect(self.remove_from_user_dict)
-        buttons_layout.addWidget(self.btn_remove_dict, 2, 1)
+heading=ImageTk.PhotoImage(Image.open(r"/home/Suchitra/Desktop/code./dictionary/title_dark.jpeg"))
+mainlabel=tk.Label(image=heading)
+mainlabel.place(x=0,y=0)
 
-        self.btn_display_history = QPushButton("Display History", self)
-        self.set_button_style(self.btn_display_history)
-        self.btn_display_history.clicked.connect(self.display_history)
-        buttons_layout.addWidget(self.btn_display_history, 3, 0)
+label1=tk.Label(root,text="Enter text here:",font=("georgia",24))
+label1.place(x=2,y=200)
 
-        self.clear_btn = QPushButton("Clear", self)
-        self.set_button_style(self.clear_btn)
-        self.clear_btn.clicked.connect(self.clear_output)
-        buttons_layout.addWidget(self.clear_btn, 3, 1)
+entry1=tk.Entry(font="georgia",fg="black")
+entry1.place(x=240,y=210)
 
-        layout.addLayout(buttons_layout)
+b1=customtkinter.CTkButton(master=root,text="spellcheck",command=spellcheck)
+b1.place(x=440,y=210)
 
-    def set_button_style(self, button):
-        button.setFont(QFont("Arial", 20))
-        button.setFixedHeight(70)
-        button.setStyleSheet("""
-            QPushButton {
-                border-radius: 20px;
-                padding: 15px;
-                background-color: #2E7D32;  
-                color: white;
-            }
-            QPushButton:hover {
-                background-color: #66bb6a;  
-            }
-            QPushButton:pressed {
-                background-color: #388e3c;  
-            }
-        """)
+b2=tk.Button(root,text="Meanings and Origin",bg="pink",fg="black",font="georgia",borderwidth=10,command=get_word_info)
+b2.place(x=2,y=260)
 
-    def create_output_area(self, layout):
-        self.output_area = QTextEdit(self)
-        self.output_area.setReadOnly(True)
-        self.output_area.setFont(QFont("Arial", 18))
-        self.output_area.setStyleSheet("""
-            QTextEdit {
-                background-color: #2A2A2A;  
-                border-radius: 20px;
-                padding: 15px;
-                border: 2px solid #FFEA00;  
-                color: white;
-            }
-        """)
-        layout.addWidget(self.output_area)
+b3=tk.Button(root,text="Synonyms",bg="pink",fg="black",font="georgia",borderwidth=10,command=get_synonyms)
+b3.place(x=200,y=260)
 
-    def spellcheck(self):
-        text = self.text_input.text()
-        history_tracking.append(text)
-        blob = TextBlob(text)
-        corrected_text = str(blob.correct())
+b4=tk.Button(root,text="My Dictionary",bg="pink",fg="black",font="georgia",borderwidth=10,command=disp_ud)
+b4.place(x=315,y=260)
 
-        if corrected_text != text:
-            question = f"Did you mean: '{corrected_text}'?"
-            reply = QMessageBox.question(self, 'Spell Check', question, QMessageBox.StandardButton.Yes | QMessageBox.StandardButton.No)
-            
-            if reply == QMessageBox.StandardButton.Yes:
-                self.text_input.setText(corrected_text)
-            else:
-                self.output_area.setText("No better suggestions were found.")
-        else:
-            self.output_area.setText("No spelling errors found.")
+b5=tk.Button(root,text="Add to my dictionary",bg="pink",fg="black",font="georgia",borderwidth=10,command=add_to_user_dict)
+b5.place(x=450,y=260)
 
-    def pronounce_word(self):
-        word = self.text_input.text()
-        eng.say(word)
-        eng.runAndWait()
+b6=tk.Button(root,text="Remove from my Dictionary",bg="pink",fg="black",font="georgia",borderwidth=10,command=remove_from_dic)
+b6.place(x=2,y=320)
 
-    def get_word_info(self):
-        word = self.text_input.text()
-        synsets = wordnet.synsets(word)
-        meanings = [syn.definition() for syn in synsets] if synsets else []
-        origins = self.get_word_origin(word)
+b7=tk.Button(root,text="My History",bg="pink",fg="black",font="georgia",borderwidth=10,command=disp_history)
+b7.place(x=250,y=320)
 
-        output = f"Word: {word}\nMeanings: {meanings}\nOrigin: {origins}"
-        self.output_area.setText(output)
 
-    def get_word_origin(self, word):
-        synsets = wordnet.synsets(word)
-        origins = set()
-        for syn in synsets:
-            for lemma in syn.lemmas():
-                origins.add(lemma.name())
-        return origins
+b8=tk.Button(root,image=speaker,command=pronun)
+#b8=tk.Button(root,text="Pronounciation",bg="pink",fg="black",font="georgia",borderwidth=10,command=pronun)
+b8.place(x=640,y=220)
 
-    def get_synonyms(self):
-        word = self.text_input.text()
-        synonyms = set()
+b9=tk.Button(root,text="Clear",bg="pink",fg="black",font="georgia",borderwidth=10,command=remove)
+b9.place(x=370,y=320)
 
-        for syn in wordnet.synsets(word):
-            for lemma in syn.lemmas():
-                synonyms.add(lemma.name())
+mode_button = tk.Button(root, text="Light Mode",bg="pink",fg="black",font="georgia",borderwidth=10, command=toggle_mode)
+mode_button.place(x=510,y=320)
 
-        if synonyms:
-            self.output_area.setText(f"Synonyms of '{word}':\n" + ", ".join(synonyms))
-        else:
-            self.output_area.setText(f"No synonyms found for '{word}'.")
 
-    def add_to_user_dict(self):
-        word = self.text_input.text()
-        user_dict.add(word)
-        self.output_area.setText(f"'{word}' added to My Dictionary.")
 
-    def remove_from_user_dict(self):
-        word = self.text_input.text()
-        user_dict.discard(word)
 
-    def display_user_dict(self):
-        if user_dict:
-            self.output_area.setText(f"My Dictionary:\n" + ", ".join(user_dict))
-        else:
-            self.output_area.setText("My Dictionary is empty.")
-
-    def display_history(self):
-        if history_tracking:
-            self.output_area.setText(f"History:\n" + ", ".join(history_tracking))
-        else:
-            self.output_area.setText("History is empty.")
-
-    def clear_output(self):
-        self.output_area.setText("")
-
-if __name__ == '__main__':
-    app = QApplication(sys.argv)
-    window = PhraseCraftApp()
-    window.show()
-    sys.exit(app.exec())
+root.mainloop()
